+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/PriorityClassTest.java
@@ -16,10 +16,10 @@
 
 package io.fabric8.kubernetes.client.mock;
 
-import io.fabric8.kubernetes.api.model.scheduling.PriorityClass;
-import io.fabric8.kubernetes.api.model.scheduling.PriorityClassBuilder;
-import io.fabric8.kubernetes.api.model.scheduling.PriorityClassList;
-import io.fabric8.kubernetes.api.model.scheduling.PriorityClassListBuilder;
+import io.fabric8.kubernetes.api.model.scheduling.v1beta1.PriorityClass;
+import io.fabric8.kubernetes.api.model.scheduling.v1beta1.PriorityClassBuilder;
+import io.fabric8.kubernetes.api.model.scheduling.v1beta1.PriorityClassList;
+import io.fabric8.kubernetes.api.model.scheduling.v1beta1.PriorityClassListBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.server.mock.KubernetesServer;